refactor(rbr): merge Dispute + Evacuation actors into single RuleBasedActor
One actor per peer; each tick runs the dispute branch then the evacuation
branch, each branch short-circuits on the treasury state of the other.
Unified Event + Error ADTs. RBRM hands it both loaders and a single tracer.
