+++ src/test/resources/logback.xml
@@ -22,8 +22,7 @@
     <logger name="JointLedger" level="warn"/>
     <logger name="HeadMultisigRegimeManager" level="warn"/>
     <logger name="CoilMultisigRegimeManager" level="warn"/>
-    <logger name="DisputeActor" level="warn"/>
-    <logger name="EvacuationActor" level="warn"/>
+    <logger name="RuleBasedActor" level="warn"/>
     <logger name="Limiter" level="warn"/>
     <logger name="Persistence" level="warn"/>
     <logger name="CoilPeerWsTransport" level="warn"/>
